# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "All HTTP + MQ endpoints exposed by the goop2 viewer.\\n\\nAll peer-to-peer signaling travels through MQ (POST /api/mq/send → P2P → /api/mq/events SSE).\\nSee x-mq-topics in the raw spec (/api/openapi.json) for the full topic contract.",
        "title": "goop2 Viewer API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/groups": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List hosted groups with live member data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/routes.hostedGroupInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Create a new hosted group",
                "parameters": [
                    {
                        "description": "Create request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.groupCreateResponse"
                        }
                    }
                }
            }
        },
        "/api/groups/close": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Close and delete a hosted group (broadcasts group:close via MQ to all members)",
                "parameters": [
                    {
                        "description": "Group ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/invite": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Invite a peer to a hosted group (sends group.invite via MQ)",
                "parameters": [
                    {
                        "description": "Invite request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/join": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Join a remote group as a member (sends group:join via MQ)",
                "parameters": [
                    {
                        "description": "Join request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupHostJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/join-own": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Host joins own group as a member",
                "parameters": [
                    {
                        "description": "Group ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/kick": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Kick a member from a hosted group",
                "parameters": [
                    {
                        "description": "Kick request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/leave": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Leave a group as a member (sends group:leave via MQ)",
                "parameters": [
                    {
                        "description": "Group ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/leave-own": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Host leaves own group",
                "parameters": [
                    {
                        "description": "Group ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/max-members": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Update max member limit for a hosted group",
                "parameters": [
                    {
                        "description": "Max members request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupMaxMembersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/meta": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Update group name and/or max_members (broadcasts group:meta via MQ)",
                "parameters": [
                    {
                        "description": "Meta request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupMetaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/rejoin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Rejoin a previously joined group",
                "parameters": [
                    {
                        "description": "Rejoin request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupHostJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
        "/api/groups/send": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Send a payload to a group (host broadcasts, member sends to host)",
                "parameters": [
                    {
                        "description": "Send request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupSendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
...
        "/api/groups/subscriptions/remove": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Remove a stale subscription record",
                "parameters": [
                    {
                        "description": "Remove request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.groupHostJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routes.statusOK"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "routes.groupCreateRequest": {
            "type": "object",
            "properties": {
                "group_context": {
                    "type": "string"
                },
                "group_type": {
                    "type": "string"
                },
                "max_members": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "My Group"
                },
                "volatile": {
                    "type": "boolean"
                }
            }
        },
        "routes.groupCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "status": {
                    "type": "string",
                    "example": "created"
                }
            }
        },
        "routes.groupHostJoinRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "host_peer_id": {
                    "type": "string",
                    "example": "12D3KooWXxx..."
                }
            }
        },
        "routes.groupIDRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                }
            }
        },
        "routes.groupMaxMembersRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "max_members": {
                    "type": "integer",
                    "example": 20
                }
            }
        },
        "routes.groupMemberInfo": {
            "type": "object",
            "properties": {
                "joined_at": {
                    "type": "integer",
                    "example": 1709136000000
                },
                "name": {
                    "type": "string",
                    "example": "Roadwarrior"
                },
                "peer_id": {
                    "type": "string",
                    "example": "12D3KooWXxx..."
                }
            }
        },
        "routes.groupMetaRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "max_members": {
                    "type": "integer",
                    "example": 20
                },
                "name": {
                    "type": "string",
                    "example": "New Name"
                }
            }
        },
        "routes.groupPeerRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "peer_id": {
                    "type": "string",
                    "example": "12D3KooWXxx..."
                }
            }
        },
        "routes.groupSendRequest": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "payload": {}
            }
        },
        "routes.hostedGroupInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-03-08T12:00:00Z"
                },
                "group_context": {
                    "type": "string",
                    "example": "Friday Jams"
                },
                "group_type": {
                    "type": "string",
                    "example": "listen"
                },
                "host_in_group": {
                    "type": "boolean"
                },
                "host_joined": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6a1b2"
                },
                "max_members": {
                    "type": "integer",
                    "example": 20
                },
                "member_count": {
                    "type": "integer",
                    "example": 3
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/routes.groupMemberInfo"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "My Group"
                },
                "volatile": {
                    "type": "boolean"
                }
            }
        },
...
        "routes.statusOK": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    }
}